// packages/doppler-v4-indexer/ponder.schema.ts
export const asset = onchainTable(
  "asset",
  (t) => ({
    address: t.hex().primaryKey(),
    poolId: t.hex().notNull(),
    isToken0: t.boolean().notNull(),
    poolAddress: t.hex().notNull(),
    chainId: t.bigint().notNull(),
    numeraire: t.hex().notNull(),
    timelock: t.hex().notNull(),
    governance: t.hex().notNull(),
    liquidityMigrator: t.hex().notNull(),
    poolInitializer: t.hex().notNull(),
    migrationPool: t.hex().notNull(),
    v2Pool: t.hex(),
    numTokensToSell: t.bigint().notNull(),
    integrator: t.hex().notNull(),
    createdAt: t.bigint().notNull(),
    migratedAt: t.bigint(),
    migrated: t.boolean().notNull().default(false),
    percentDayChange: t.real().notNull().default(0),
    marketCapUsd: t.bigint().notNull().default(0n),
    holderCount: t.integer().notNull().default(0),
    dayVolumeUsd: t.bigint().notNull().default(0n),
    liquidityUsd: t.bigint().notNull().default(0n),
  }),
  (table) => ({
    addressIdx: index().on(table.address),
    chainIdIdx: index().on(table.chainId),
  })
);
// ...
export const v2Pool = onchainTable("v2_pool", (t) => ({
  poolId: t.hex().notNull().primaryKey(),
  address: t.hex().notNull(),
  chainId: t.bigint().notNull(),
  baseToken: t.hex().notNull(),
  quoteToken: t.hex().notNull(),
  reserveBaseToken: t.bigint().notNull(),
  reserveQuoteToken: t.bigint().notNull(),
  totalFeeBaseToken: t.bigint().notNull(),
  totalFeeQuoteToken: t.bigint().notNull(),
  price: t.bigint().notNull(),
  parentPool: t.hex().notNull(),
  v3Pool: t.hex().notNull(),
  migratedAt: t.bigint(),
  migrated: t.boolean().notNull(),
  isToken0: t.boolean().notNull(),
}));